requirements_a00.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
loguru==0.5.1
path==14.0.1
psutil==5.7.0
pygame==1.9.6
requests==2.24.0
urllib3==1.25.9
urlpath==1.1.7
xdgappdirs==1.4.5
